requirements.txt
torch
tqdm.auto
torchvision
matplotlib.pyplot